refactor(dashboard): split the time-series sections out of views.py (#1105 Phase 3 D4, develop-v2)
Cut D4 of the dashboard cluster: the sections that carry a history rather than a current
reading — share health (#116), the Tier-1 telemetry backbone's blocks/disk-growth/XvB-history
formatters (#196), the pool/mode palette tokens (#27) and the hashrate and cadence cards —
move from web/views.py into a new web/series_views.py. Their test classes move 1:1.

Pure move. The boundary comes from the forward dependency closure, not the cut map's line
ranges: 17 top-level names, one contiguous run plus the 4-line palette block, with zero
retained names interleaved. Everything leaves the closure only to build_state, so views.py
stays the facade; _gauge_series is re-exported unused for web/worker_detail.py, and
_filter_events becomes a pure re-export for the same reason.

views.py 660 -> 445, tests/web/test_views.py 1228 -> 861.
