+++ docs/dev/file-budget.tsv
@@ -20,7 +20,7 @@ dashboard/mining_dashboard/web/static/logic.mjs	534
 dashboard/mining_dashboard/web/static/osupdate.mjs	415
 dashboard/mining_dashboard/web/static/wizard.mjs	889
 dashboard/mining_dashboard/web/static/workerview.mjs	443
-dashboard/mining_dashboard/web/views.py	660
+dashboard/mining_dashboard/web/views.py	445
 dashboard/mining_dashboard/web/xvb_views.py	827
 dashboard/mining_dashboard/wizard.py	674
 dashboard/tests/client/test_xmrig_client.py	504
@@ -40,8 +40,9 @@ dashboard/tests/service/test_metrics.py	463
 dashboard/tests/service/test_storage_service.py	1646
 dashboard/tests/service/test_telegram_commands.py	1240
 dashboard/tests/web/test_infra_views.py	865
+dashboard/tests/web/test_series_views.py	484
 dashboard/tests/web/test_server.py	1319
-dashboard/tests/web/test_views.py	1228
+dashboard/tests/web/test_views.py	861
 dashboard/tests/web/test_wizard.py	974
 dashboard/tests/web/test_xvb_views.py	1342
 .github/workflows/ci.yml	513
